requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Configuration and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.5.0

# Caching and storage
redis>=5.0.1

# Security
argon2-cffi>=23.1.0

# Observability
prometheus-client>=0.19.0
structlog>=24.1.0
